RTSP: fix Media player wrong indication on Starting a Live Stream.
There will show 1:00:00 total duration  on visible mumeric
cunter of Media Player. Since the Live Streaming will not
finish, so there is a wrong display information.
From the Android website:
https://developer.android.com/reference/android/media
/MediaPlayer.html
There is information about API int getDuration(),which
said that the duration in milliseconds, if no duration is
available(for example,if streaming live content) ,-1 is
returned.

Test: 1. Go go Live Streaming website;
      2. Click on start for playing;
      3. Start a Live streaming session;
      4. Check the total time counter.

Change-Id: Ia16686ffa1257a6c932a65b14867e7198ecef477